Display speciality entity in display statistics. Get data about submission from metadavalue field instead of parsing annotation.

+++ dspace-jspui/src/main/java/ua/edu/sumdu/essuir/utils/EssuirUtils.java
@@ -188,7 +188,7 @@ public class EssuirUtils {
         return findAuthor(surname, initials);
     }
 
-    public static Map<Speciality, Integer> getSpecialityStatistics(LocalDate from, LocalDate to) {
+    public static Map<Speciality, Long> getSpecialityStatistics(LocalDate from, LocalDate to) {
         return specialityStatisticsService.getSpecialityStatistics(from, to);
     }
